// src/app/core/models/phenotype-query-api.interface.ts
import { IAql } from './aql.interface'
import { LogicalOperator } from './logical-operator.enum'
import { PhenotypeQueryType } from './phenotype-query-type.enum'

export interface IPhenotypeQueryApi {
  type: PhenotypeQueryType
  operator?: LogicalOperator
  children?: IPhenotypeQueryApi[]
  aql?: IAql
}